+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -1,0 +1,16 @@
+//
+//  Generated file. Do not edit.
+//
+
+import FlutterMacOS
+import Foundation
+
+import device_info_plus_macos
+import motor_flutter
+import path_provider_macos
+
+func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  DeviceInfoPlusMacosPlugin.register(with: registry.registrar(forPlugin: "DeviceInfoPlusMacosPlugin"))
+  MotorFlutterPlugin.register(with: registry.registrar(forPlugin: "MotorFlutterPlugin"))
+  PathProviderPlugin.register(with: registry.registrar(forPlugin: "PathProviderPlugin"))
+}